MoneyOps/ai-gateway/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
groq>=0.4.1
anthropic>=0.8.1
python-dotenv>=1.0.0
tenacity>=8.2.3
structlog>=24.1.0
redis>=5.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.2.1
livekit-api>=0.1.0
livekit-agents[groq,silero,openai]>=0.8.0
rapidfuzz>=3.6.0


# Dev dependencies
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-mock>=3.12.0
pytest-cov>=4.1.0
black>=23.12.1
ruff>=0.1.9
mypy>=1.8.0
